fix subaccount balance query v2

+++ lib/src/rest/classes.dart
@@ -563,8 +563,8 @@ class SubaccountBalance {
 
   SubaccountBalance.fromMap(Map m)
       : asset = m['asset'],
-        free = m['free'],
-        locked = m['locked'];
+        free = m['free'] as double,
+        locked = m['locked'] as double;
 }
 
 class SubaccountBalances {
